+++ saleor/attribute/models/page.py
@@ -1,3 +1,4 @@
+from django.contrib.postgres.indexes import BTreeIndex
 from django.db import models
 
 from ...core.models import SortableModel
@@ -16,10 +17,19 @@ class AssignedPageAttributeValue(SortableModel):
         on_delete=models.CASCADE,
         related_name="pagevalueassignment",
     )
+    page = models.ForeignKey(
+        Page,
+        related_name="attributevalues",
+        on_delete=models.CASCADE,
+        null=True,
+        blank=True,
+        db_index=False,
+    )
 
     class Meta:
         unique_together = (("value", "assignment"),)
         ordering = ("sort_order", "pk")
+        indexes = [BTreeIndex(fields=["page"], name="assignedpageattrvalue_page_idx")]
 
     def get_ordering_queryset(self):
         return self.assignment.pagevalueassignment.all()